test(openclaw): isolate MCP inventory coverage
